Fixed: A bug in child bitmap creation caused excessive memory consumption and as a consequence could result in program crashes Fixed: An ambiguous stack template class call caused problems in sound handling Fixed: A bug in canvas handling could cause program crashes
git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@12605 01821946-851b-0410-b7f4-f1477c70e57a

+++ audio/audio.cpp
@@ -914,7 +914,7 @@ return m_usedChannels.Push (int (channelP - m_channels.Buffer ())) ? m_usedChann
 void CAudio::UnregisterChannel (int nIndex)
 {
 if ((nIndex >= 0) && (nIndex < int (m_usedChannels.ToS ()))) {
-	m_usedChannels.Delete (nIndex);
+	m_usedChannels.Delete ((uint) nIndex);
 	if (nIndex < int (m_usedChannels.ToS ())) {
 #if 0 //DBG
 		if (!m_channels [m_usedChannels [nIndex]].Playing ())
